+++ tenon/packages/global.d.ts
@@ -1,2 +1,4 @@
 declare var __GLOBAL__: any
 declare var __DEV__: Boolean
+declare const Hummer: any
+declare const Memory:any;
